+++ src/Blamite/Blam/SecondGen/MetaOffsetConverter.cs
@@ -13,22 +13,22 @@ namespace Blamite.Blam.SecondGen
 			_mask = mask;
 		}
 
-		public int PointerToOffset(uint pointer)
+		public int PointerToOffset(long pointer)
 		{
 			return PointerToOffset(pointer, _metaSegment.Offset);
 		}
 
-		public int PointerToOffset(uint pointer, int areaStartOffset)
+		public int PointerToOffset(long pointer, int areaStartOffset)
 		{
 			return (int) (pointer - _mask + areaStartOffset);
 		}
 
-		public uint OffsetToPointer(int offset)
+		public long OffsetToPointer(int offset)
 		{
 			return OffsetToPointer(offset, _metaSegment.Offset);
 		}
 
-		public uint OffsetToPointer(int offset, int areaStartOffset)
+		public long OffsetToPointer(int offset, int areaStartOffset)
 		{
 			return (uint) (offset - areaStartOffset + _mask);
 		}
