+++ src/Blamite/Blam/ThirdGen/MetaAddressConverter.cs
@@ -26,38 +26,38 @@ namespace Blamite.Blam.ThirdGen
 	public class MetaAddressConverter : IPointerConverter
 	{
 		private readonly FileSegment _metaSegment;
-		private uint _virtualBase;
+		private long _virtualBase;
 
 		/// <summary>
 		///     Constructs a new MetaAddressConverter.
 		/// </summary>
 		/// <param name="metaSegment">The FileSegment where meta is stored.</param>
 		/// <param name="virtualBase">The virtual base address of the meta.</param>
-		public MetaAddressConverter(FileSegment metaSegment, uint virtualBase)
+		public MetaAddressConverter(FileSegment metaSegment, long virtualBase)
 		{
 			_metaSegment = metaSegment;
 			_virtualBase = virtualBase;
 			metaSegment.Resized += MetaResized;
 		}
 
-		public int PointerToOffset(uint pointer)
+		public int PointerToOffset(long pointer)
 		{
 			return PointerToOffset(pointer, _metaSegment.Offset);
 		}
 
-		public int PointerToOffset(uint pointer, int areaStartOffset)
+		public int PointerToOffset(long pointer, int areaStartOffset)
 		{
 			return (int) (pointer - _virtualBase + areaStartOffset);
 		}
 
-		public uint OffsetToPointer(int offset)
+		public long OffsetToPointer(int offset)
 		{
 			return OffsetToPointer(offset, _metaSegment.Offset);
 		}
 
-		public uint OffsetToPointer(int offset, int areaStartOffset)
+		public long OffsetToPointer(int offset, int areaStartOffset)
 		{
-			return (uint) (offset - areaStartOffset + _virtualBase);
+			return (offset - areaStartOffset + _virtualBase);
 		}
 
 		private void MetaResized(object sender, SegmentResizedEventArgs e)
