+++ src/Blamite/Blam/ThirdGen/Resources/BSP/ThirdGenScenarioBSP.cs
@@ -35,7 +35,7 @@ namespace Blamite.Blam.ThirdGen.Resources.BSP
 			EngineDescription buildInfo)
 		{
 			var count = (int) values.GetInteger("number of sections");
-			uint address = values.GetInteger("section table address");
+			uint address = (uint)values.GetInteger("section table address");
 			StructureLayout layout = buildInfo.Layouts.GetLayout("model section");
 			StructureValueCollection[] entries = ReflexiveReader.ReadReflexive(reader, count, address, layout, metaArea);
 
@@ -47,7 +47,7 @@ namespace Blamite.Blam.ThirdGen.Resources.BSP
 			EngineDescription buildInfo)
 		{
 			var count = (int) values.GetInteger("number of bounding boxes");
-			uint address = values.GetInteger("bounding box table address");
+			uint address = (uint)values.GetInteger("bounding box table address");
 			StructureLayout layout = buildInfo.Layouts.GetLayout("model bounding box");
 			StructureValueCollection[] entries = ReflexiveReader.ReadReflexive(reader, 1, address, layout, metaArea);
