+++ src/Blamite/Blam/ThirdGen/Resources/Models/ThirdGenRenderModel.cs
@@ -36,7 +36,7 @@ namespace Blamite.Blam.ThirdGen.Resources.Models
 			EngineDescription buildInfo)
 		{
 			var count = (int) values.GetInteger("number of regions");
-			uint address = values.GetInteger("region table address");
+			uint address = (uint)values.GetInteger("region table address");
 			StructureLayout layout = buildInfo.Layouts.GetLayout("model region");
 			StructureValueCollection[] entries = ReflexiveReader.ReadReflexive(reader, count, address, layout, metaArea);
 
@@ -48,7 +48,7 @@ namespace Blamite.Blam.ThirdGen.Resources.Models
 			EngineDescription buildInfo)
 		{
 			var count = (int) values.GetInteger("number of sections");
-			uint address = values.GetInteger("section table address");
+			uint address = (uint)values.GetInteger("section table address");
 			StructureLayout layout = buildInfo.Layouts.GetLayout("model section");
 			StructureValueCollection[] entries = ReflexiveReader.ReadReflexive(reader, count, address, layout, metaArea);
 
@@ -60,7 +60,7 @@ namespace Blamite.Blam.ThirdGen.Resources.Models
 			EngineDescription buildInfo)
 		{
 			var count = (int) values.GetInteger("number of bounding boxes");
-			uint address = values.GetInteger("bounding box table address");
+			uint address = (uint)values.GetInteger("bounding box table address");
 			StructureLayout layout = buildInfo.Layouts.GetLayout("model bounding box");
 			StructureValueCollection[] entries = ReflexiveReader.ReadReflexive(reader, 1, address, layout, metaArea);
