+++ src/Blamite/Util/XMLUtil.cs
@@ -16,11 +16,11 @@ namespace Blamite.Util
 		/// <param name="numberStr">The string to parse. If it begins with "0x", it will be parsed as hexadecimal.</param>
 		/// <param name="result">The variable to store the parsed value to if parsing succeeds.</param>
 		/// <returns>true if parsing the string succeeded.</returns>
-		public static bool ParseNumber(string numberStr, out int result)
+		public static bool ParseNumber(string numberStr, out long result)
 		{
 			if (numberStr.StartsWith("0x"))
-				return int.TryParse(numberStr.Substring(2), NumberStyles.HexNumber, null, out result);
-			return int.TryParse(numberStr, out result);
+				return long.TryParse(numberStr.Substring(2), NumberStyles.HexNumber, null, out result);
+			return long.TryParse(numberStr, out result);
 		}
 
 		/// <summary>
@@ -34,14 +34,14 @@ namespace Blamite.Util
 		/// <exception cref="ArgumentException">Thrown if the attribute is missing.</exception>
 		/// <exception cref="FormatException">Thrown if the attribute does not represent an integer.</exception>
 		/// <seealso cref="ParseNumber" />
-		public static int GetNumericAttribute(XElement element, string name)
+		public static long GetNumericAttribute(XElement element, string name)
 		{
 			XAttribute attribute = element.Attribute(name);
 			if (attribute == null)
 				throw new ArgumentException("A(n) \"" + element.Name + "\" element is missing the required \"" + name +
 											"\" attribute.");
 
-			int result;
+			long result;
 			if (ParseNumber(attribute.Value, out result))
 				return result;
 
@@ -59,13 +59,13 @@ namespace Blamite.Util
 		/// <returns>The attribute's value, or the default value if the attribute was not found.</returns>
 		/// <exception cref="FormatException">Thrown if the attribute does not represent an integer.</exception>
 		/// <seealso cref="ParseNumber" />
-		public static int GetNumericAttribute(XElement element, string name, int defaultValue)
+		public static long GetNumericAttribute(XElement element, string name, int defaultValue)
 		{
 			XAttribute attribute = element.Attribute(name);
 			if (attribute == null)
 				return defaultValue;
 
-			int result;
+			long result;
 			if (ParseNumber(attribute.Value, out result))
 				return result;
